backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyyaml==6.0.1
httpx==0.26.0
python-dateutil==2.8.2
alembic==1.13.1
mcstatus==11.1.1
aiosqlite==0.19.0
mcstatus==11.1.1
